Finalized model training, evaluation, and tuning.
I have finalized the model training, evaluation, and tuning. This and more code has been completed but am working on cleaning up to make the code presentable. Currently testing a live implementation of the project which has been exciting, but that code needs some cleaning up before being added to the project. I am very satisfied with the prediction accuracy of the model, up from previous projects I have done, and being more thoughtful with feature engineering seems to have paid off. I still, however haven't iteratively improved the model, and likely have gone overboard with the number of features at this stage. Several features are correlated and dimensionality could probably be reduced. That is on the list of improvements. Anyhow, here is a long overdue update to the project as it stands.